feat: Allow sorting of types when rendering a template
Since Python dicts remember the order of insertions (as of Python 3.7), this commit updates the code to ensure the commit types are in a certain, user-configurable, order when being rendered in a template.

Release-Notes: The new "type_order" configuration key allows users to specify the order commit types are dprovided in when rendering a template.